Updated c++ demos for win32
git-svn-id: http://svn.code.sf.net/p/glscene/code@6634 c452ef9c-9f09-11df-acb8-e9cdb1ccbceb

+++ trunk/Samples/CPP/Demos/materials/mirror/Unit1.cpp
@@ -120,7 +120,7 @@ void __fastcall TForm1::FormMouseWheel(TObject *Sender, TShiftState Shift, int W
 		  TPoint &MousePos, bool &Handled)
 {
   GLCamera1->
-   AdjustDistanceToTarget(Power(1.1, (float)((float)WheelDelta / 120.0)));
+   AdjustDistanceToTarget(Power(1.1, (WheelDelta / 120.0)));
 }
 //---------------------------------------------------------------------------
