+++ trunk/Samples/CPP/Demos/sprites/imposter/Unit1.cpp
@@ -1,6 +1,8 @@
 //---------------------------------------------------------------------------
 
 #include <vcl.h>
+#include <tchar.h>
+
 #pragma hdrstop
 
 #include "Unit1.h"
@@ -126,7 +128,7 @@ void __fastcall TForm1::FormMouseWheel(TObject *Sender, TShiftState Shift, int W
           TPoint &MousePos, bool &Handled)
 {
   GLCamera1->
-   AdjustDistanceToTarget(Power(1.1, (float)((float)WheelDelta / 120.0)));
+   AdjustDistanceToTarget(Power(1.1, (WheelDelta / 120.0)));
 }
 //---------------------------------------------------------------------------
